back to github

+++ 0x0F-function_pointers/2-int_index.c
@@ -1,7 +1,7 @@
 #include "function_pointers.h"
 
 /**
-* array_iterator -  description
+* int_index -  description
 *
 * @array: input function argument
 * @size: input function argument
@@ -13,15 +13,19 @@
 int int_index(int *array, int size, int (*cmp)(int))
 
 {
-size_t i;
+int i;
 
-for (i=0; i < size; i++)
-while (cmp(array[i]) > 0){
-if (i == size){
+for (i = 0; i < size; i++)
+while (cmp(array[i]) > 0)
+{
+if (i == size)
+{
 printf("%d", size);
 return (-1);
 }
-else {
-return (i);}
+else
+{
+return (i);
+}
 }
 }
